[#8] 필요없는 노드 삭제 및 플레이어 감지 서비스 노드 추가

+++ Source/Y2024Q1_Prototyping/Public/Entity/Character/Base/LLL_BaseCharacter.h
@@ -50,7 +50,7 @@ protected:
 	uint32 CharacterMoveSpeed;
 
 	UPROPERTY(VisibleAnywhere)
-	uint32 bIsDead : 1;
+	uint8 bIsDead : 1;
 
 #if (WITH_EDITOR || UE_BUILD_DEVELOPMENT)
 	// 디버그용 함수
